Exportar Excel
- Listo, con los datos obtenidos desde la Base de datos.

+++ JAVA/Entidades/src/com/fshoes/entidades/Orden.java
@@ -64,7 +64,7 @@ public class Orden {
         this.fecha_entrega = fecha_entrega;
     }    
 
-    public float getTotal() {
+    public int getTotal() {
         return total;
     }
